Render Glyphs only once
When rendering Text, the glyphs will be rendered by FreeType only once
per glyph, and then stored on the graphics card. The glyphs are then
kept in a GlyphLibrary.

With this, we will no longer need to create and store multiple
textures of the same glyphs.

The next step is to actually render a Text to screen using gfx-rs and
OpenGL.